+++ packages/nimble-components/src/table/components/cell/index.ts
@@ -8,7 +8,7 @@ import {
 } from '@microsoft/fast-element';
 import { DesignSystem, FoundationElement } from '@microsoft/fast-foundation';
 import type { MenuButtonToggleEventDetail } from '../../../menu-button/types';
-import type { TableCellRecord, TableCellState } from '../../types';
+import type { TableCellRecord, TableCellState } from '../../../table-column/base/types';
 import { styles } from './styles';
 import { template } from './template';
