+++ packages/nimble-components/src/table/components/cell/tests/table-cell.pageobject.ts
@@ -1,5 +1,5 @@
 import type { TableCell } from '..';
-import type { TableCellRecord } from '../../../types';
+import type { TableCellRecord } from '../../../../table-column/base/types';
 
 /**
  * Page object for the `nimble-table-cell` component to provide consistent ways
