+++ packages/nimble-components/src/table/components/row/index.ts
@@ -3,10 +3,12 @@ import { DesignSystem, FoundationElement } from '@microsoft/fast-foundation';
 import { styles } from './styles';
 import { template } from './template';
 import type {
+    TableCellState
+} from '../../../table-column/base/types';
+import type {
     TableActionMenuToggleEventDetail,
-    TableCellState,
-    TableDataRecord,
-    TableFieldName
+    TableFieldName,
+    TableRecord
 } from '../../types';
 import type { TableColumn } from '../../../table-column/base';
 import type { MenuButtonToggleEventDetail } from '../../../menu-button/types';
@@ -21,6 +23,10 @@ export interface ColumnState {
     column: TableColumn;
     cellState: TableCellState;
 }
+
+/** Represents a single row (element) in the Table's data  */
+// eslint-disable-next-line @typescript-eslint/no-empty-interface
+interface TableDataRecord extends TableRecord {}
 
 /**
  * A styled row that is used within the nimble-table.
