+++ packages/nimble-components/src/table/tests/types.spec.ts
@@ -1,6 +1,5 @@
 import type {
-    TableColumnSortDirection,
-    TableColumnSortOperation
+    TableColumnSortDirection
 } from '../types';
 
 describe('Table type', () => {
@@ -8,11 +7,5 @@ describe('Table type', () => {
         // @ts-expect-error This expect will fail if the enum-like type is missing "as const"
         const sortDirection: TableColumnSortDirection = 'hello';
         expect(sortDirection!).toEqual('hello');
-    });
-
-    it('TableColumnSortOperation fails compile if assigning arbitrary string values', () => {
-        // @ts-expect-error This expect will fail if the enum-like type is missing "as const"
-        const sortOperation: TableColumnSortOperation = 'hello';
-        expect(sortOperation).toEqual('hello');
     });
 });
